servidor/rest/requirements.txt
fastapi
uvicorn
pymongo
python-jose
jsonschema
